Fix duplicate style properties in college components and update college-company features
