--- .env
@@ -1,38 +1,0 @@
-# MongoDB Atlas Connection String (with URL-encoded password)
-MONGODB_URI=mongodb+srv://harsh:harsh123@campus.k0by8i8.mongodb.net/
-# MONGODB_URI=mongodb+srv://harsh:harsh123@campus.k0by8i8.mongodb.net/campus?retryWrites=true&w=majority
-# Port your backend server will run on
-PORT=5000
-
-# Secret key for express-session or JWT
-SESSION_SECRET=SuperSecretSessionKey
-
-# Sales and Support 
-AWS_ACCESS_KEY_ID=your_aws_access_key 
-AWS_SECRET_ACCESS_KEY=your_aws_secret_key 
-AWS_REGION=your_aws_region 
-S3_BUCKET_NAME=your_s3_bucket_name 
-
-DIGILOCKER_BASE_URL=https://api.digilocker.gov.in
-DIGILOCKER_CLIENT_ID=your_digilocker_client_id 
-DIGILOCKER_CLIENT_SECRET=your_digilocker_client_secret 
-DIGILOCKER_REDIRECT_URI=http://localhost:3001/digilocker/callback
-
-
-# College & Company
-#Zoom Meetings Scheduling
-
-ZOOM_ACCOUNT_ID=JuUgCUOQRyWACmgqMjLeuQ
-ZOOM_CLIENT_ID=OMaVaVUqSu6DYBU8et4knQ
-ZOOM_CLIENT_SECRET=438Cd2Ua39BiHHwgAh2eqgJnrHxmwFiz
-ZOOM_USER_ID=harsh@rishishwarindustry.in
-
-#Sending Emails
-EMAIL_SENDER=hr.snowflake.melissa@gmail.com
-EMAIL_PASS=agengdvmdbcnhanc
-
-REDIS_HOST=localhost
-REDIS_PORT=6379
-
-#BACKEND URL
-VITE_API_URL=
