+++ Frontend/.env
@@ -1,1 +1,1 @@
-VITE_API_URL=https://campusadmin.onrender.com
+VITE_API_URL=http://localhost:5000
