+++ Frontend/src/college_components/CollegeLogin.jsx
@@ -152,7 +152,7 @@ const CollegeLogin = () => {
     setEmailCheckLoading(true);
     try {
       // Send all registration data (except password) to initiate endpoint
-      const res = await axios.post(`${apiUrl}/api/college/register/initiate`, {
+      const res = await axios.post(`${apiUrl}/api/colleges/register/initiate`, {
         name: registerData.name,
         code: registerData.code,
         location: registerData.location,
@@ -204,7 +204,7 @@ const CollegeLogin = () => {
 
     setRegisterLoading(true);
     try {
-      await axios.post(`${apiUrl}/api/college/register/verify`, {
+      const res = await axios.post(`${apiUrl}/api/colleges/register/verify`, {
         email: contactEmail,
         otp: otp,
         password,
@@ -243,7 +243,7 @@ const CollegeLogin = () => {
 
     try {
       // Send minimal data required by initiate endpoint to resend OTP
-      await axios.post(`${apiUrl}/api/college/register/initiate`, {
+      const res = await axios.post(`${apiUrl}/api/colleges/register/initiate`, {
         contactEmail: registerData.contactEmail,
         // Include other necessary identification if your backend initiate endpoint requires it, like type
         type: 'college', // Ensure type is sent
@@ -303,7 +303,7 @@ const CollegeLogin = () => {
 
     try {
       // Call the new backend endpoint to check OTP validity without completing registration
-      const res = await axios.post(`${apiUrl}/api/register/check-otp`, {
+      const res = await axios.post(`${apiUrl}/api/auth/register/check-otp`, {
         email: registerData.contactEmail,
         otp: otp,
         type: 'college' // Specify the type
