+++ backend/.env
@@ -1,5 +1,6 @@
 # MongoDB Atlas Connection String (with URL-encoded password)
-MONGODB_URI=mongodb+srv://harsh:harsh123@campus.k0by8i8.mongodb.net/
+#MONGODB_URI=mongodb+srv://harsh:harsh123@campus.k0by8i8.mongodb.net/
+MONGODB_URI=mongodb://localhost:27017/campus_connect
 # MONGODB_URI=mongodb+srv://harsh:harsh123@campus.k0by8i8.mongodb.net/campus?retryWrites=true&w=majority
 # Port your backend server will run on
 PORT=5000
