+++ backend/config/email.js
@@ -1,0 +1,10 @@
+const nodemailer = require('nodemailer');
+module.exports.emailTransport = nodemailer.createTransport({
+  host: 'smtp.gmail.com',
+  port: 587,
+  secure: false,
+  auth: {
+    user: process.env.EMAIL_SENDER,
+    pass: process.env.EMAIL_PASS
+  }
+});
